requirements.txt
# ViewGuard Student Monitor - 고정확도 졸음 감지 시스템
# Python 3.8 이상 필요

# 핵심 라이브러리
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
Pillow==10.1.0

# 자동화
pyautogui==0.9.54

# 알림 시스템
python-telegram-bot==20.7

# 구글 시트 연동 (선택)
gspread==5.12.0
oauth2client==4.1.3

# 웹훅 / API
requests==2.31.0

# 유틸리티
python-dotenv==1.0.0
